+++ app/helpers/ask/monitoring/dashboard_helper.rb
@@ -1,0 +1,9 @@
+module Ask
+  module Monitoring
+    module DashboardHelper
+      def metric_chart_data(metric)
+        metric.as_chart_data.to_json
+      end
+    end
+  end
+end